feat: 불필요한 generation 반환 로직 변경 (#324)

+++ src/main/java/org/sopt/app/application/soptamp/SoptampUserInfo.java
@@ -30,7 +30,7 @@ public class SoptampUserInfo {
         private Long userId;
         private Long playgroundId;
         private String name;
-        private Long generation;
+        private Integer generation;
         private String part;
     }
