+++ src/main/java/org/sopt/app/domain/entity/SoptampUser.java
@@ -38,7 +38,7 @@ public class SoptampUser extends BaseEntity {
     private String nickname;
 
     @Column(name = "generation")
-    private Long generation;
+    private Integer generation;
 
     @Column(name = "part")
     private String part;
@@ -51,7 +51,7 @@ public class SoptampUser extends BaseEntity {
         this.nickname = nickname;
     }
 
-    public void updateGenerationAndPart(Long generation, PlaygroundPart part) {
+    public void updateGenerationAndPart(Integer generation, PlaygroundPart part) {
         this.generation = generation;
         this.part = part.name();
     }
